firebase/backend/requirements.txt
google-cloud-firestore
mock-firestore